Fixed issues with field/property info not reading correctly. Now working on card's having a null game instance and throwing an invocation exception on acting value being called.

+++ Testing/ConsoleApplication1/Gwent/GwentInstance/AI/OpponentBaseUtils.cs
@@ -53,7 +53,7 @@ public abstract class OpponentBaseUtils : Player
             return new Move(Card.Clone, Type, RowType);
         }
         
-        public int  GetAiHash()
+        public long GetAiHash()
         {
             return AiUtils.HashInternalElements(this);
         }
