+++ Testing/ConsoleApplication1/Gwent/GwentInstance/Player.cs
@@ -106,7 +106,7 @@ public class Player : IHashable
         return output.ToString();
     }
 
-    public int GetAiHash()
+    public long GetAiHash()
     {
         return AiUtils.HashInternalElements(this);
     }
